pyproject.toml
[project]
name = "UltraStitch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "h5py>=3.15.1",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "py-mon==2.0.5",
    "pywebview>=6.1",
    "requests>=2.32.5",
    "scipy>=1.16.3",
]
